refactor(uri): remove legacy code and move test utilities
- Remove parseLocation, formatLocation, and location conversion helpers
- Remove deprecated ParsedURIResult type
- Move InMemorySymbolIndex from resolver.ts to test helpers
- Update index.ts exports to reflect removals
- Update tests to import InMemorySymbolIndex from test helper
- Add overload disambiguation documentation to ADR-0044
