+++ packages/devac-core/src/uri/index.ts
@@ -42,7 +42,6 @@ export type {
   EntityID,
   Location,
   ParsedURI,
-  ParsedURIResult,
   SymbolIndex,
   SymbolIndexEntry,
   URIQueryParams,
@@ -67,13 +66,10 @@ export {
   isEntityID,
   isRelativeRef as isRelativeRefString,
   isSymbolPath,
-  locationToQueryParams,
   parseCanonicalURI,
   parseEntityID,
-  parseLocation,
   parseQueryParams,
   parseSymbolPath,
-  queryParamsToLocation,
   URIParseError,
 } from "./parser.js";
 
@@ -86,7 +82,6 @@ export {
   createSymbolPath,
   formatCanonicalURI,
   formatEntityID,
-  formatLocation,
   formatQueryParams,
   formatSymbolPath,
   formatSymbolSegment,
@@ -102,7 +97,6 @@ export {
   getParentURI,
   getURIDepth,
   getURIFromEntityID,
-  InMemorySymbolIndex,
   resolveURIToEntityID,
   symbolPathsEqual,
   uriMatchesContext,
